Housekeeping: tiny changes.

+++ zz/src/internal/vulkan/render.h
@@ -17,7 +17,9 @@
 #include "sync.h"
 #include "buffer.h"
 
-typedef struct {u32* data; u16 length, capacity;} zz_memory_array_u16_t;
+#include "manifest.h"
+
+typedef struct {u16* data; u16 length, capacity;} zz_memory_array_u16_t;
 
 struct zz_internal_render
 {
